# Make small perf refactors, clippy and grammar fixes